Move BiomesCommand and BiomesSetCommand to user command package.
